Allow creation of and access to pending Subscriptions
Honor 'type' attributes on complex elements. The value will be the nodename of
some resource class. This properly supports the API's <pending_subscription
type="subscription"> formation.

Also fully reset instances after _update(), so the subscription instances are
reset to the values we get back in the response. This resets, for example,
'timeframe' back to unset and 'unit_amount_in_cents' back to the subscription's
current value, even though it had to be locally set to the future value to
create the pending subscription.