Support unicode characters in resource elements.
The current policy is to cast values to str when serializing them to XML
elements. If a unicode string is received (which is likely to happen in
people names) a unicode encode error will be raised.
The current commit ensures the value is casted to unicode, which is safe
for ascii strings as well. The tests fixtures haven been updated to
consider unicode characters in values.