Purchases endpoint
Added the purchases endpoint. Also changed the serialization of lists.
We no longer render the attribute `type="array"` because the server does
not need it and it was causing a problem on the purchases endpoint.
There is a TODO in this code to look at removing the currency default
feature as sometimes embedding resources requires the currency to be
stated at the top level (such as in the case of purchases).