V2 add processing_prepayment_balance_in_cents to AccountBalance
Add new add processing_prepayment_balance_in_cents attribute to
AccountBalance. The processing_prepayment_balance_in_cents attribute
is a similar format to the balance_in_cents attribute and contains
the total for processing prepayment credit invoices in each currency.
This value is useful when trying to determine if the customer's
prepayment balance has run out or if it is currently processing
while waiting on a payment for the corresponding purchase invoice.